パスワードをメールアドレス認証後に入力するように変更
